fix: use Integer ids for purchase requisition entities

+++ mmssystem/src/main/java/com/g174/mmssystem/entity/PurchaseRequisition.java
@@ -21,7 +21,7 @@ public class PurchaseRequisition {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "requisition_id")
-    private Long requisitionId;
+    private Integer requisitionId;
 
     @Column(name = "requisition_no", length = 30, unique = true, nullable = false)
     private String requisitionNo;
